fix(cilium): k3s has an API server proxy at `127.0.0.1:6444`
There is no need to use an external LB or something like kube-vip for this Cilium config, k3s has an API server proxy listening in `127.0.0.1:6444` on all nodes in the cluster. I ran with this setup in the https://github.com/onedr0p/cluster-template back when using k3s so just sharing some knowledge to here.